applied fixes
