+++ app/shared/components/design-system/all-components/breadcrumbs/BreadCrumbsClient.tsx
@@ -1,10 +1,12 @@
 'use client';
 
-import * as React from 'react';
-import { usePathname } from 'next/navigation';
 import { Box, Typography } from '@mui/material';
-import CustomLink from '../CustomLink/CustomLink';
+import { usePathname } from 'next/navigation';
+import * as React from 'react';
+
 import { BreadCrumbsStyles } from './BreadCrumbs.styles';
+
+import CustomLink from '~/shared/components/design-system/all-components/link/CustomLink';
 
 const BreadcrumbsClient = () => {
   const pathname = usePathname();
